DCMAW-11524 | Refactor scripts for clean up and change endpoints to be the existing buckets
