--- backend/src/mockData/BitStringStatusListResponseJWT
@@ -1,1 +1,0 @@
-eyJhbGciOiJFUzI1NiIsImtpZCI6IjEyIiwidHlwIjoidmMrand0In0.eyJleHAiOjIyOTE3MjAxNzAsImlhdCI6MTY4NjkyMDE3MCwic3RhdHVzX2xpc3QiOnsiYml0cyI6MiwibHN0IjoiZUp3ek1BQUNReVFNSkFFbmxnUEYifSwic3ViIjoiaHR0cHM6Ly9jcnMuYWNjb3VudC5nb3YudWsvdC8zQjBGM0JEMDg3QTciLCJ0dGwiOjQzMjAwfQ.S3PNiQFjC3V_c7In_eLNY2-Nv2AWs0M2m3NpXsJqN7haPwQewbJenT41DH_I5gU0RV4klfKVMPtHUL-cz8DB1A
