--- backend/src/mockData/TokenStatusListResponseCWT
@@ -1,1 +1,0 @@
-0oRDoQEmogN4GmFwcGxpY2F0aW9uL3N0YXR1c2xpc3QrY3d0BEIxMaVjZXhwGoiY3-pjaWF0GmSMW-pjc3VieClodHRwczovL2Nycy5hY2NvdW50Lmdvdi51ay90LzNCMEYzQkQwODdBN2N0dGwZqMBrc3RhdHVzX2xpc3SiY2xzdHRlSnd6TUFBQ1F5UU1KQUVubGdQRmRiaXRzAlhAgYqqCZbWQgxdmIEwvZL3yzQh7BSi_qEHHQab1OyVNzO3Gk1GFn8z2cyASDbvXByU1sLAIpnVwR-GH78LLAgtEQ
